build(config): :art: change structure to export components with small bundles
