+++ packages/core/src/CoreExample/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const CoreExample = () => {
+  return <section>Core Example! :)</section>;
+};
+
+export default CoreExample;
