--- packages/core/src/components/CoreExample/coreExample.module.css
@@ -1,3 +1,0 @@
-.coreTest {
-  color: red;
-}
