--- packages/core/src/components/CoreExample/index.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './coreExample.module.css';
-
-const CoreExample = () => {
-  return <div className={`${styles.coreTest}`}>Core Example! :)</div>;
-};
-
-export default CoreExample;
