--- packages/core/src/components/index.ts
@@ -1,3 +1,0 @@
-import CoreExample from './coreExample';
-
-export { CoreExample };
