+++ packages/home/src/HomeExample/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const HomeExample = () => {
+  return <section>Ejemplo del home! :)</section>;
+};
+
+export default HomeExample;
