--- packages/home/src/components/homeExample/homeExample.module.css
@@ -1,3 +1,0 @@
-.homeTest {
-  color: green;
-}
