--- packages/home/src/components/homeExample/index.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './homeExample.module.css';
-
-const HomeExample = () => {
-  return <div className={`${styles.homeTest}`}>Ejemplo del home! :)</div>;
-};
-
-export default HomeExample;
