--- packages/home/src/components/index.ts
@@ -1,3 +1,0 @@
-import HomeExample from './homeExample';
-
-export { HomeExample };
