+++ packages/product/src/ProductExample/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ProductExample = () => {
+  return <section>Este es un ejemplo de Product! :)</section>;
+};
+
+export default ProductExample;
