--- packages/product/src/components/index.ts
@@ -1,3 +1,0 @@
-import ProductExample from './productExample';
-
-export { ProductExample };
