--- packages/product/src/components/productExample/index.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './productExample.module.css';
-
-const ProductExample = () => {
-  return <div className={`${styles.productTest}`}>Este es un ejemplo de Product! :)</div>;
-};
-
-export default ProductExample;
