--- packages/product/src/components/productExample/productExample.module.css
@@ -1,3 +1,0 @@
-.productTest {
-  color: blue;
-}
