+++ packages/search/src/SearchExample/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const SearchExample = () => {
+  return <section>Este es un ejemplo del Search! :)</section>;
+};
+
+export default SearchExample;
