--- packages/search/src/components/index.ts
@@ -1,3 +1,0 @@
-import SearchExample from './searchExample';
-
-export { SearchExample };
