--- packages/search/src/components/searchExample/index.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './searchExample.module.css';
-
-const SearchExample = () => {
-  return <div className={`${styles.searchTest}`}>Este es un ejemplo del Search! :)</div>;
-};
-
-export default SearchExample;
