--- packages/search/src/components/searchExample/searchExample.module.css
@@ -1,3 +1,0 @@
-.searchTest {
-  color: yellow;
-}
